Dashing - good | Camera when moving down and left - fix

+++ Assets/Scripts/Enemy/EnemyController.cs
@@ -31,7 +31,7 @@ public class EnemyController : MonoBehaviour
       health--;
     }
 
-    public void TakeDamage(int d)
+    public void TakeDamage(float d)
     {
       health -= d;
     }
